add add_stacker
